refactor(tests): extract fixed_point_u4x12_u8x8.cpp from fixed_point.cpp (part of #3130)
Fourth extraction in #3130. Moves the u4x12 + u8x8 unsigned type tests
into their own file. Both share a section because they share the
file-local approx_equal() helper (used only within these two sections).

Content moved:
  - 28 FL_TEST_CASE u4x12 tests (construction, arithmetic, math ops,
    edge cases, precision, compound operations)
  - 60+ FL_TEST_CASE u8x8 tests (construction variants, addition/sub/
    mul/div, shifts, comparisons, math ops, sqrt/rsqrt/pow, edge cases)
  - approx_equal() static helper (lines 1644-1646 of original)

Tests pass: bash test fl_math_fixed_point --cpp (1/1, 0.23s test time,
44s build).

fixed_point.cpp: 4,277 -> 3,404 LOC
fixed_point_u4x12_u8x8.cpp (new): 896 LOC

Stacked on feat/3130-fixed-point-bounds-header (PR #3138). Predecessors:
  - feat/3130-fixed-point-misc-extract (PR #3135 — hook + misc)
  - feat/3130-fixed-point-constexpr-extract (PR #3136 — constexpr)
  - feat/3130-fixed-point-bounds-header (PR #3138 — bounds.h)

Part of #3130. Meta tracking issue: #3127.
